💅 Attempt to install prettier packages, too (#4)
* 💅 Attempt to intall prettier packages, too

We're seeing some weird prettier errors when linting a project that has eslint configured to run prettier. My theory is that different prettier versions are being installed depending on peer dependencies vs top-level dependencies. 🤞

* Try to nstall babel dependencies, too

The same strategy that was used for eslint, applied to babel. Parsing has
an impact on linting, so I'm curious to see if this makes the remaining
3 errors on my example branch go away.

* Revert "Try to nstall babel dependencies, too"

This reverts commit 2df54b0f0b3470abba93bd379d87bae70358106d.