Allow dependencies to be defined
This commit allows a balto-eslint user to specify which packages they
need installed. The problem this solves is that, for example, if you use
`@babel/eslint-parser`, then that requires `@babel/core`. This is fine,
and it will get installed because it is a peer dependency. However, now
your babel config comes into play, and any plugins that are defined
there are required as well. However, those won't have been installed at
any point.

This commit also moves "prettier" into that status. In theory, there's
no reason we should assume the user wants that installed.