docs: New version process updates
Changelog
---

In other balto-* repos we are leaning into using Github release to
generate changelogs and have had good results with it. Cutting a new
release becomes much easier and can be done almost entirely in the
Github UI now. Because of this, we should make note of it in our
CHANGELOG.

Readme
---

The semantic tag versioner creates the following tags from `v0.8.0`

- `v0`
- `v0.8`

Normally, I wouldn't feel great about having `v0` on the readme. But I
am expecting that we will have a major version bump (`v1.0.0`) for this
repo within the next few weeks as we bring over the check run ->
workflow command changes we've already added to balto-rubocop.