feat!: Use workflow output commands instead of check runs
This solves the annoying check-run-mislabled issue, but is a breaking
change: When an action run has annotations to create, by default, it
will now mark the job as successful instead of neutral.

More explanation, discussion, context can be found on this commit
message. We hope to "unbreak" this once Github (re)adds the behavior we
need.

https://github.com/planningcenter/balto-rubocop/pull/32/commits/7b5adc8f0ad667ec5ae0ca85c19608e789e99a7a