fix: Ignore irrelevant warnings/errors in counts
eslint lints by files, but we only care about _specific lines_ in each
file. If you made a "clean" change in a "dirty" file, we would report
the total number of eslint errors and warnings found — even if we didn't
generate any annotations because your diff was good.

This change has us adding up errors and warning from the changes we deem
relevant to the PR instead of using the original numbers from eslint.

Fixes #3