feat(Debug): add logging
I needed to add this to help track down a bug, and I think it would
likely be beneficial if we left it here. There is a non-zero performance
cost to doing this, but I'm hopeful that it's small enough to be worth
the cost for when this is necessary.

I also needed to add the @actions/core package to get the debug logging
working.