Use extensions input to filter files passed to ESLint
Without this manual filtering, ESLint will attempt to lint files that
should be ignored (like ruby files). ESLint v9 has the `files` option in
the flat config, and earlier versions have the `--ext` CLI flag. Let's
aim for maximum compatibility and just ignore the files that don't end
in the extensions we want.